Fix HTML/YML server crash issue (#988)
