+++ src/client/PortalWebView.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from "vscode";
 import * as path from "path";
-import { searchPortalConfigFolder } from "../common/PortalConfigFinder";
+import { searchPortalConfigFolder } from "../common/utilities/PathFinderUtil";
 
 /**
  * Displays Portal html webpage preview
@@ -130,7 +130,7 @@ export class PortalWebView {
         if (uri) {
             // Add bootstrap.min.css
             let url = webview.asWebviewUri(
-                vscode.Uri.joinPath( uri as vscode.Uri, "web-files", "bootstrap.min.css")
+                vscode.Uri.joinPath(uri as vscode.Uri, "web-files", "bootstrap.min.css")
             );
             const bootstrap = `<link href="${url}" rel="stylesheet" />`;
             html += bootstrap;
