+++ src/common/utilities/PathFinderUtil.ts
@@ -10,8 +10,6 @@ import { URL } from 'url';
 import * as path from 'path';
 import * as fs from 'fs';
 import { glob } from 'glob';
-import { sendTelemetryEvent} from "../client/power-pages/telemetry";
-import { ITelemetry } from "../client/telemetry/ITelemetry";
 
 const portalConfigFolderName = '.portalconfig';
 
@@ -71,31 +69,4 @@ function isSibling(file: string): URL | null {
         }
     }
     return null;
-}
-
-export function getPortalsOrgURLs(workspaceRootFolders:WorkspaceFolder[] | null, telemetry: ITelemetry ) {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let output: any[] = [];
-    try {
-        workspaceRootFolders?.forEach(workspaceRootFolder => {
-            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-            const manifestFiles  = glob.sync('**/*-manifest.yml', { dot: true, cwd: workspaceRootFolder!.uri });
-            if (manifestFiles.length == 0) {
-                output = [{
-                    orgURL: '',
-                    isManifestExists: false
-                }];
-            }else {
-                manifestFiles?.forEach(manifestFile =>{
-                    output.push({
-                        orgURL: manifestFile.split("-manifest")[0].replace(/.*[portalconfig]\//,''),
-                        isManifestExists: true
-                    });
-                })
-            }
-        });
-    }catch(exception){
-        sendTelemetryEvent(telemetry, { methodName:getPortalsOrgURLs.name,eventName: 'getPortalsOrgURLs', exception: exception as Error });
-    }
-    return output;
 }
